Moving _DatastoreAPI implementations into dedicated module.
