+++ unit_tests/datastore/test_connection.py
@@ -14,286 +14,6 @@
 
 import unittest
 
-from google.cloud.datastore.connection import _HAVE_GRPC
-
-
-class Test_DatastoreAPIOverHttp(unittest.TestCase):
-
-    def _getTargetClass(self):
-        from google.cloud.datastore.connection import _DatastoreAPIOverHttp
-        return _DatastoreAPIOverHttp
-
-    def _makeOne(self, *args, **kw):
-        return self._getTargetClass()(*args, **kw)
-
-    def test__rpc(self):
-        class ReqPB(object):
-
-            def SerializeToString(self):
-                return REQPB
-
-        class RspPB(object):
-
-            def __init__(self, pb):
-                self._pb = pb
-
-            @classmethod
-            def FromString(cls, pb):
-                return cls(pb)
-
-        REQPB = b'REQPB'
-        PROJECT = 'PROJECT'
-        METHOD = 'METHOD'
-        URI = 'http://api-url'
-        conn = _Connection(URI)
-        datastore_api = self._makeOne(conn)
-        http = conn.http = Http({'status': '200'}, 'CONTENT')
-        response = datastore_api._rpc(PROJECT, METHOD, ReqPB(), RspPB)
-        self.assertTrue(isinstance(response, RspPB))
-        self.assertEqual(response._pb, 'CONTENT')
-        called_with = http._called_with
-        self.assertEqual(called_with['uri'], URI)
-        self.assertEqual(called_with['method'], 'POST')
-        self.assertEqual(called_with['headers']['Content-Type'],
-                         'application/x-protobuf')
-        self.assertEqual(called_with['headers']['User-Agent'],
-                         conn.USER_AGENT)
-        self.assertEqual(called_with['body'], REQPB)
-        self.assertEqual(conn.build_kwargs,
-                         [{'method': METHOD, 'project': PROJECT}])
-
-    def test__request_w_200(self):
-        PROJECT = 'PROJECT'
-        METHOD = 'METHOD'
-        DATA = b'DATA'
-        URI = 'http://api-url'
-        conn = _Connection(URI)
-        datastore_api = self._makeOne(conn)
-        http = conn.http = Http({'status': '200'}, 'CONTENT')
-        self.assertEqual(datastore_api._request(PROJECT, METHOD, DATA),
-                         'CONTENT')
-        called_with = http._called_with
-        self.assertEqual(called_with['uri'], URI)
-        self.assertEqual(called_with['method'], 'POST')
-        self.assertEqual(called_with['headers']['Content-Type'],
-                         'application/x-protobuf')
-        self.assertEqual(called_with['headers']['User-Agent'],
-                         conn.USER_AGENT)
-        self.assertEqual(called_with['body'], DATA)
-        self.assertEqual(conn.build_kwargs,
-                         [{'method': METHOD, 'project': PROJECT}])
-
-    def test__request_not_200(self):
-        from google.cloud.exceptions import BadRequest
-        from google.rpc import status_pb2
-
-        error = status_pb2.Status()
-        error.message = 'Entity value is indexed.'
-        error.code = 9  # FAILED_PRECONDITION
-
-        PROJECT = 'PROJECT'
-        METHOD = 'METHOD'
-        DATA = 'DATA'
-        URI = 'http://api-url'
-        conn = _Connection(URI)
-        datastore_api = self._makeOne(conn)
-        conn.http = Http({'status': '400'}, error.SerializeToString())
-        with self.assertRaises(BadRequest) as exc:
-            datastore_api._request(PROJECT, METHOD, DATA)
-        expected_message = '400 Entity value is indexed.'
-        self.assertEqual(str(exc.exception), expected_message)
-        self.assertEqual(conn.build_kwargs,
-                         [{'method': METHOD, 'project': PROJECT}])
-
-
-class Test_DatastoreAPIOverGRPC(unittest.TestCase):
-
-    def _getTargetClass(self):
-        from google.cloud.datastore.connection import _DatastoreAPIOverGRPC
-        return _DatastoreAPIOverGRPC
-
-    def _makeOne(self, stub, connection=None, secure=True, mock_args=None):
-        from unit_tests._testing import _Monkey
-        from google.cloud.datastore import connection as MUT
-
-        if connection is None:
-            connection = _Connection(None)
-            connection.credentials = object()
-            connection.host = 'CURR_HOST'
-
-        if mock_args is None:
-            mock_args = []
-
-        def mock_make_stub(*args):
-            mock_args.append(args)
-            return stub
-
-        if secure:
-            to_monkey = {'make_secure_stub': mock_make_stub}
-        else:
-            to_monkey = {'make_insecure_stub': mock_make_stub}
-        with _Monkey(MUT, **to_monkey):
-            return self._getTargetClass()(connection, secure)
-
-    def test_constructor(self):
-        from google.cloud.datastore import connection as MUT
-
-        conn = _Connection(None)
-        conn.credentials = object()
-        conn.host = 'CURR_HOST'
-
-        stub = _GRPCStub()
-        mock_args = []
-        datastore_api = self._makeOne(stub, connection=conn,
-                                      mock_args=mock_args)
-        self.assertIs(datastore_api._stub, stub)
-
-        self.assertEqual(mock_args, [(
-            conn.credentials,
-            conn.USER_AGENT,
-            MUT.datastore_grpc_pb2.DatastoreStub,
-            conn.host,
-        )])
-
-    def test_constructor_insecure(self):
-        from google.cloud.datastore import connection as MUT
-
-        conn = _Connection(None)
-        conn.credentials = object()
-        conn.host = 'CURR_HOST:1234'
-
-        stub = _GRPCStub()
-        mock_args = []
-        datastore_api = self._makeOne(stub, connection=conn,
-                                      secure=False,
-                                      mock_args=mock_args)
-        self.assertIs(datastore_api._stub, stub)
-
-        self.assertEqual(mock_args, [(
-            MUT.datastore_grpc_pb2.DatastoreStub,
-            conn.host,
-        )])
-
-    def test_lookup(self):
-        return_val = object()
-        stub = _GRPCStub(return_val)
-        datastore_api = self._makeOne(stub=stub)
-
-        request_pb = _RequestPB()
-        project = 'PROJECT'
-        result = datastore_api.lookup(project, request_pb)
-        self.assertIs(result, return_val)
-        self.assertEqual(request_pb.project_id, project)
-        self.assertEqual(stub.method_calls,
-                         [(request_pb, 'Lookup')])
-
-    def test_run_query(self):
-        return_val = object()
-        stub = _GRPCStub(return_val)
-        datastore_api = self._makeOne(stub=stub)
-
-        request_pb = _RequestPB()
-        project = 'PROJECT'
-        result = datastore_api.run_query(project, request_pb)
-        self.assertIs(result, return_val)
-        self.assertEqual(request_pb.project_id, project)
-        self.assertEqual(stub.method_calls,
-                         [(request_pb, 'RunQuery')])
-
-    def test_begin_transaction(self):
-        return_val = object()
-        stub = _GRPCStub(return_val)
-        datastore_api = self._makeOne(stub=stub)
-
-        request_pb = _RequestPB()
-        project = 'PROJECT'
-        result = datastore_api.begin_transaction(project, request_pb)
-        self.assertIs(result, return_val)
-        self.assertEqual(request_pb.project_id, project)
-        self.assertEqual(
-            stub.method_calls,
-            [(request_pb, 'BeginTransaction')])
-
-    def test_commit_success(self):
-        return_val = object()
-        stub = _GRPCStub(return_val)
-        datastore_api = self._makeOne(stub=stub)
-
-        request_pb = _RequestPB()
-        project = 'PROJECT'
-        result = datastore_api.commit(project, request_pb)
-        self.assertIs(result, return_val)
-        self.assertEqual(request_pb.project_id, project)
-        self.assertEqual(stub.method_calls,
-                         [(request_pb, 'Commit')])
-
-    def _commit_failure_helper(self, exc, err_class):
-        stub = _GRPCStub(side_effect=exc)
-        datastore_api = self._makeOne(stub=stub)
-
-        request_pb = _RequestPB()
-        project = 'PROJECT'
-        with self.assertRaises(err_class):
-            datastore_api.commit(project, request_pb)
-
-        self.assertEqual(request_pb.project_id, project)
-        self.assertEqual(stub.method_calls,
-                         [(request_pb, 'Commit')])
-
-    @unittest.skipUnless(_HAVE_GRPC, 'No gRPC')
-    def test_commit_failure_aborted(self):
-        from grpc import StatusCode
-        from grpc._channel import _Rendezvous
-        from grpc._channel import _RPCState
-        from google.cloud.exceptions import Conflict
-
-        details = 'Bad things.'
-        exc_state = _RPCState((), None, None, StatusCode.ABORTED, details)
-        exc = _Rendezvous(exc_state, None, None, None)
-        self._commit_failure_helper(exc, Conflict)
-
-    @unittest.skipUnless(_HAVE_GRPC, 'No gRPC')
-    def test_commit_failure_cancelled(self):
-        from grpc import StatusCode
-        from grpc._channel import _Rendezvous
-        from grpc._channel import _RPCState
-
-        exc_state = _RPCState((), None, None, StatusCode.CANCELLED, None)
-        exc = _Rendezvous(exc_state, None, None, None)
-        self._commit_failure_helper(exc, _Rendezvous)
-
-    @unittest.skipUnless(_HAVE_GRPC, 'No gRPC')
-    def test_commit_failure_non_grpc_err(self):
-        exc = RuntimeError('Not a gRPC error')
-        self._commit_failure_helper(exc, RuntimeError)
-
-    def test_rollback(self):
-        return_val = object()
-        stub = _GRPCStub(return_val)
-        datastore_api = self._makeOne(stub=stub)
-
-        request_pb = _RequestPB()
-        project = 'PROJECT'
-        result = datastore_api.rollback(project, request_pb)
-        self.assertIs(result, return_val)
-        self.assertEqual(request_pb.project_id, project)
-        self.assertEqual(stub.method_calls,
-                         [(request_pb, 'Rollback')])
-
-    def test_allocate_ids(self):
-        return_val = object()
-        stub = _GRPCStub(return_val)
-        datastore_api = self._makeOne(stub=stub)
-
-        request_pb = _RequestPB()
-        project = 'PROJECT'
-        result = datastore_api.allocate_ids(project, request_pb)
-        self.assertIs(result, return_val)
-        self.assertEqual(request_pb.project_id, project)
-        self.assertEqual(
-            stub.method_calls,
-            [(request_pb, 'AllocateIds')])
-
 
 class TestConnection(unittest.TestCase):
 
@@ -1086,55 +806,3 @@ class Http(object):
     def request(self, **kw):
         self._called_with = kw
         return self._response, self._content
-
-
-class _Connection(object):
-
-    host = None
-    USER_AGENT = 'you-sir-age-int'
-
-    def __init__(self, api_url):
-        self.api_url = api_url
-        self.build_kwargs = []
-
-    def build_api_url(self, **kwargs):
-        self.build_kwargs.append(kwargs)
-        return self.api_url
-
-
-class _GRPCStub(object):
-
-    def __init__(self, return_val=None, side_effect=Exception):
-        self.return_val = return_val
-        self.side_effect = side_effect
-        self.method_calls = []
-
-    def _method(self, request_pb, name):
-        self.method_calls.append((request_pb, name))
-        return self.return_val
-
-    def Lookup(self, request_pb):
-        return self._method(request_pb, 'Lookup')
-
-    def RunQuery(self, request_pb):
-        return self._method(request_pb, 'RunQuery')
-
-    def BeginTransaction(self, request_pb):
-        return self._method(request_pb, 'BeginTransaction')
-
-    def Commit(self, request_pb):
-        result = self._method(request_pb, 'Commit')
-        if self.side_effect is Exception:
-            return result
-        else:
-            raise self.side_effect
-
-    def Rollback(self, request_pb):
-        return self._method(request_pb, 'Rollback')
-
-    def AllocateIds(self, request_pb):
-        return self._method(request_pb, 'AllocateIds')
-
-
-class _RequestPB(object):
-    project_id = None
